+++ source/src/main/java/com/sixgroup/refit/observability/item35/creator/domain/model/StorageCapacityDto.java
@@ -3,6 +3,8 @@ package com.sixgroup.refit.observability.item35.creator.domain.model;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
+
 @RequiredArgsConstructor
 @Data
 public class StorageCapacityDto {
@@ -10,9 +12,9 @@ public class StorageCapacityDto {
     private String reportingDate;
     private String date;
     private String timeStamp;
-    private Float capacity;
-    private Float usedCapacity;
-    private Float availableCapacity;
-    private Float utilization;
+    private BigDecimal capacity;
+    private BigDecimal usedCapacity;
+    private BigDecimal availableCapacity;
+    private BigDecimal utilization;
 
 }
